feat: implementa endpoints do back no app do angular

+++ back/atm/DTOs/Response/UserInfoDTO.cs
@@ -5,8 +5,8 @@
         public int UserId { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public int CurrentAccountBalance { get; set; }
-        public int SavingsAccountBalance { get; set; }
+        public long CurrentAccountBalance { get; set; }
+        public long SavingsAccountBalance { get; set; }
         public DateTime? LastAcces {  get; set; }
     }
 }
